LayoutTests/requestidlecallback/requestidlecallback-in-page-cache-with-pending-task-expected.txt

This tests that requestIdleCallback will execute the function even when there is a pending task for a suspended document

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS didExecuteIdleCallback is true
PASS successfullyParsed is true

TEST COMPLETE
